+++ api-augment/kusama/query.ts
@@ -14,7 +14,7 @@ export type __AugmentedQuery<ApiType extends ApiTypes> = AugmentedQuery<ApiType,
 export type __QueryableStorageEntry<ApiType extends ApiTypes> = QueryableStorageEntry<ApiType>;
 
 declare module 'https://deno.land/x/polkadot/api-base/types/storage.ts' {
-  export interface AugmentedQueries<ApiType extends ApiTypes> {
+  interface AugmentedQueries<ApiType extends ApiTypes> {
     auctions: {
       /**
        * Number of auctions started so far.