+++ api-augment/kusama/tx.ts
@@ -14,7 +14,7 @@ export type __SubmittableExtrinsic<ApiType extends ApiTypes> = SubmittableExtrin
 export type __SubmittableExtrinsicFunction<ApiType extends ApiTypes> = SubmittableExtrinsicFunction<ApiType>;
 
 declare module 'https://deno.land/x/polkadot/api-base/types/submittable.ts' {
-  export interface AugmentedSubmittables<ApiType extends ApiTypes> {
+  interface AugmentedSubmittables<ApiType extends ApiTypes> {
     auctions: {
       /**
        * Make a new bid from an account (including a parachain account) for deploying a new