+++ api-augment/substrate/consts.ts
@@ -10,7 +10,7 @@ import type { FrameSupportPalletId, FrameSupportWeightsRuntimeDbWeight, FrameSys
 export type __AugmentedConst<ApiType extends ApiTypes> = AugmentedConst<ApiType>;
 
 declare module 'https://deno.land/x/polkadot/api-base/types/consts.ts' {
-  export interface AugmentedConsts<ApiType extends ApiTypes> {
+  interface AugmentedConsts<ApiType extends ApiTypes> {
     alliance: {
       /**
        * The deposit required for submitting candidacy.