+++ api-augment/substrate/events.ts
@@ -10,7 +10,7 @@ import type { FrameSupportScheduleLookupError, FrameSupportTokensMiscBalanceStat
 export type __AugmentedEvent<ApiType extends ApiTypes> = AugmentedEvent<ApiType>;
 
 declare module 'https://deno.land/x/polkadot/api-base/types/events.ts' {
-  export interface AugmentedEvents<ApiType extends ApiTypes> {
+  interface AugmentedEvents<ApiType extends ApiTypes> {
     alliance: {
       /**
        * An ally has been elevated to Fellow.